Move tests for code coverage
