+++ hypothesis-python/tests/nocover/test_uuids.py
@@ -8,14 +8,12 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file, You can
 # obtain one at https://mozilla.org/MPL/2.0/.
 
-import uuid
 
 import pytest
 
 from hypothesis import given, strategies as st
-from hypothesis.errors import InvalidArgument
 
-from tests.common.debug import assert_no_examples, minimal
+from tests.common.debug import minimal
 
 
 @given(st.lists(st.uuids()))
@@ -35,17 +33,3 @@ def test_can_generate_specified_version(version):
         assert version == uuid.version
 
     inner()
-
-
-def test_no_nil_uuid():
-    assert_no_examples(st.uuids(), lambda x: x == uuid.UUID(int=0))
-
-
-def test_nil_uuid():
-    st.uuids(allow_nil=True), lambda x: x == uuid.UUID(int=0)
-
-
-def test_can_only_allow_nil_uuid_with_none_version():
-    st.uuids(version=None, allow_nil=True).example()
-    with pytest.raises(InvalidArgument):
-        st.uuids(version=4, allow_nil=True).example()
